# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrackType = (typeof TrackType)[keyof typeof TrackType];

export const TrackType = {
  original: "original",
  remix: "remix",
  live: "live",
  cover: "cover",
} as const;

export type TrackSource = (typeof TrackSource)[keyof typeof TrackSource];

export const TrackSource = {
  youtube: "youtube",
  soundcloud: "soundcloud",
} as const;

export interface TrackResult {
  /** Unique track ID (prefixed with source, e.g. yt_VIDEO_ID) */
  id: string;
  title: string;
  artist: string;
  type: TrackType;
  /** Duration in seconds */
  duration: number;
  source: TrackSource;
  thumbnailUrl?: string | null;
  /** Available quality options (e.g. ["128", "320"]) */
  quality: string[];
  viewCount?: number | null;
  /** Ranking score (higher is better match) */
  score: number;
}

export interface SearchRequest {
  artist: string;
  title: string;
}

export interface SearchResponse {
  /** The normalized search query used */
  query: string;
  results: TrackResult[];
  /** Whether results came from cache */
  cached: boolean;
}

export interface StreamResponse {
  id: string;
  streamUrl: string;
  mimeType?: string | null;
}

export interface DownloadResponse {
  id: string;
  downloadUrl: string;
  filename?: string | null;
}

export interface ErrorResponse {
  error: string;
  message: string;
}