# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  artist: string;
  title: string;
}