# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackResult } from "./trackResult";

export interface SearchResponse {
  /** The normalized search query used */
  query: string;
  results: TrackResult[];
  /** Whether results came from cache */
  cached: boolean;
}