# lib/api-zod/src/generated/types/trackResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackSource } from "./trackSource";
import type { TrackType } from "./trackType";

export interface TrackResult {
  /** Unique track ID (prefixed with source, e.g. yt_VIDEO_ID) */
  id: string;
  title: string;
  artist: string;
  type: TrackType;
  /** Duration in seconds */
  duration: number;
  source: TrackSource;
  thumbnailUrl?: string | null;
  /** Available quality options (e.g. ["128", "320"]) */
  quality: string[];
  viewCount?: number | null;
  /** Ranking score (higher is better match) */
  score: number;
}