# lib/api-zod/src/generated/types/trackSource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */

export type TrackSource = (typeof TrackSource)[keyof typeof TrackSource];

export const TrackSource = {
  youtube: "youtube",
  soundcloud: "soundcloud",
} as const;