# lib/api-zod/src/generated/types/trackType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */

export type TrackType = (typeof TrackType)[keyof typeof TrackType];

export const TrackType = {
  original: "original",
  remix: "remix",
  live: "live",
  cover: "cover",
} as const;